Choose and render correct final primer for Sequence Primer mode
This is an implementation.  It is brittle.  We should be passing around a full
sequence model with its relationship to the full original sequence it has
come from so that primer position and offsets are calculated correctly.
For example, at the moment, the overlap is just subtracted from the length
of the sequence the primer pair were produced against, however this length
can and will change as the overlap will not be constant.